feat(docs): clarify navigation trie functions
